Switch to standard types, replace long types
Includes `<cstdint>` and updates typedefs to point to it so that I don't
need to go around the entire source and update every single type.

Replace `long` with `int` in any place where the type is used for a
plain value. This clarifies that nothing is broken here.

Generally use `int` as the "don't care" type. `int` always means `int32_t`
on modern compilers on all platforms. This is convention.

`unsigned long` becomes `uint32_t`.

`printf` function which was converting pointers to int to print them
now uses `%p` instead to print the pointer directly.

`EffectInfo::Source` is a suspicious union of pointer and enum types.
I used `intptr_t` here for comparisons to prevent truncation.

Did not touch `long` in preprocessor or tokenizer sources.

+++ src/Base.cpp
@@ -478,7 +478,7 @@ SkipThisFree:
 
 
 
-long ZeroValue = 0;                            
+int ZeroValue = 0;
 
 template<class S,int32 Initial,int32 Delta>
   Array<S,Initial,Delta>::Array()
@@ -1256,11 +1256,11 @@ void SymbolTable::Bind(int32 id, Binding *b)
 #define UPPER_MASK 0x80000000UL /* most significant w-r bits */
 #define LOWER_MASK 0x7fffffffUL /* least significant r bits */
 
-static unsigned long mt[N]; /* the array for the state vector  */
+static uint32_t mt[N]; /* the array for the state vector  */
 static int mti=N+1; /* mti==N+1 means mt[N] is not initialized */
 
 /* initializes mt[N] with a seed */
-void init_genrand(unsigned long s)
+void init_genrand(uint32_t s)
 {
     mt[0]= s & 0xffffffffUL;
     for (mti=1; mti<N; mti++) {
@@ -1279,7 +1279,7 @@ void init_genrand(unsigned long s)
 /* init_key is the array for initializing keys */
 /* KY_length is its length */
 /* slight change for C++, 2004/2/26 */
-void init_by_array(unsigned long init_key[], int KY_length)
+void init_by_array(uint32_t init_key[], int KY_length)
 {
     int i, j, k;
     init_genrand(19650218UL);
@@ -1305,10 +1305,10 @@ void init_by_array(unsigned long init_key[], int KY_length)
 }
 
 /* generates a random number on [0,0xffffffff]-interval */
-unsigned long genrand_int32(void)
+uint32_t genrand_int32(void)
 {
-    unsigned long y;
-    static unsigned long mag01[2]={0x0UL, MATRIX_A};
+    uint32_t y;
+    static uint32_t mag01[2]={0x0UL, MATRIX_A};
     /* mag01[x] = x * MATRIX_A  for x=0,1 */
 
     if (mti >= N) { /* generate N words at one time */
@@ -1343,9 +1343,9 @@ unsigned long genrand_int32(void)
 }
 
 /* generates a random number on [0,0x7fffffff]-interval */
-long genrand_int31(void)
+int genrand_int31(void)
 {
-    return (long)(genrand_int32()>>1);
+    return (int)(genrand_int32()>>1);
 }
 
 /* generates a random number on [0,1]-real-interval */
@@ -1372,7 +1372,7 @@ double genrand_real3(void)
 /* generates a random number on [0,1) with 53-bit resolution*/
 double genrand_res53(void) 
 { 
-    unsigned long a=genrand_int32()>>5, b=genrand_int32()>>6; 
+    uint32_t a=genrand_int32()>>5, b=genrand_int32()>>6;
     return(a*67108864.0+b)*(1.0/9007199254740992.0); 
 }
